🎨 style: refactor code to remove unused imports and classes This commit removes unused imports and classes to improve code readability and maintainability.
✨ feat(reviewer_submit_gallery.dart): add ReviewerSubmitGallery widget
This commit adds a new widget called ReviewerSubmitGallery. This widget is responsible for displaying a gallery of documents and images related to a debitur. It also includes buttons to view the gallery and documents, as well as a checkbox to indicate that the user has viewed the documents.
