+++ lib/app/modules/reviewer_submit/views/reviewer_submit_view.dart
@@ -5,6 +5,7 @@ import 'dart:convert';
 import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_agunan.dart';
 import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_analys_response.dart';
 import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_bisnis.dart';
+import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_gallery.dart';
 import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_inputan.dart';
 import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_karakter.dart';
 import 'package:akm/app/modules/reviewer_submit/widget/reviewer_submit_keuangan.dart';
@@ -17,14 +18,12 @@ import 'package:form_builder_extra_fields/form_builder_extra_fields.dart';
 import 'package:form_builder_validators/form_builder_validators.dart';
 import 'package:get/get.dart';
 import 'package:getwidget/getwidget.dart';
-import 'package:google_fonts/google_fonts.dart';
 import 'package:http/http.dart' as http;
 import 'package:intl/intl.dart';
 
 // 🌎 Project imports:
 import 'package:akm/app/common/constant.dart';
 import 'package:akm/app/common/style.dart';
-import 'package:akm/app/routes/app_pages.dart';
 import '../controllers/reviewer_submit_controller.dart';
 
 // ignore: must_be_immutable
@@ -321,7 +320,9 @@ class ReviewerSubmitView extends GetView<ReviewerSubmitController> {
                           ),
                     ),
                     const SizedBox(height: 20),
-                    GalleryCard(controller: controller),
+                    ReviewerSubmitGallery(
+                      controller: controller,
+                    ),
                     const SizedBox(height: 20),
                     Container(
                       color: Colors.grey[200],
@@ -526,152 +527,6 @@ class ReviewerSubmitView extends GetView<ReviewerSubmitController> {
                 ),
               ),
             ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class GalleryCard extends StatelessWidget {
-  const GalleryCard({
-    Key? key,
-    required this.controller,
-  }) : super(key: key);
-
-  final ReviewerSubmitController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      color: Colors.grey[200],
-      child: Card(
-        child: Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: Column(
-            children: [
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  const Expanded(
-                    flex: 6,
-                    child: GFTypography(
-                      text: 'Gallery Debitur',
-                      type: GFTypographyType.typo3,
-                      showDivider: false,
-                    ),
-                  ),
-                  Expanded(
-                    child: Obx(
-                      () => Icon(
-                        controller.isGalleryRead.value == true
-                            ? Icons.check_box
-                            : Icons.close,
-                        color: controller.isGalleryRead.value == true
-                            ? Colors.green
-                            : Colors.transparent,
-                        size: 30,
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 10),
-              Text(
-                'Ini merupakan sekumpulan dokumen penunjang seperti foto dan dokumen lainnya yang sudah diunggah oleh tim analis',
-                style: Theme.of(context).textTheme.bodySmall?.merge(
-                      const TextStyle(
-                        fontSize: 14,
-                      ),
-                    ),
-              ),
-              const SizedBox(height: 20),
-              GFButton(
-                onPressed: () {
-                  Get.toNamed(Routes.GALLERY_IMAGE,
-                      arguments: controller.insightDebitur.value);
-                },
-                text: 'Lihat Gallery',
-                color: primaryColor,
-                shape: GFButtonShape.pills,
-                icon: const Icon(
-                  Icons.image,
-                  size: 18,
-                  color: secondaryColor,
-                ),
-                fullWidthButton: true,
-                size: GFSize.LARGE,
-              ),
-              const SizedBox(height: 10),
-              SizedBox(
-                width: 200,
-                child: Row(
-                  children: [
-                    const Expanded(
-                      child: Divider(
-                        color: Colors.black,
-                        thickness: 1,
-                      ),
-                    ),
-                    Text(
-                      "     ATAU     ",
-                      style: TextStyle(
-                        color: Colors.black,
-                        fontSize: 15,
-                        fontWeight: FontWeight.w800,
-                        // add custom google font
-                        fontFamily: GoogleFonts.montserrat().fontFamily,
-                      ),
-                    ),
-                    const Expanded(
-                      child: Divider(
-                        color: Colors.black,
-                        thickness: 1,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              const SizedBox(height: 10),
-              GFButton(
-                onPressed: () {
-                  Get.toNamed(Routes.GALLERY_FILE,
-                      arguments: controller.insightDebitur.value);
-                },
-                color: primaryColor,
-                shape: GFButtonShape.pills,
-                text: 'Lihat Dokumen',
-                icon: const Icon(
-                  Icons.picture_as_pdf,
-                  size: 18,
-                  color: secondaryColor,
-                ),
-                fullWidthButton: true,
-                size: GFSize.LARGE,
-              ),
-              const SizedBox(height: 10),
-              FormBuilderCheckbox(
-                name: 'berkas',
-                onChanged: (value) {
-                  controller.isGalleryRead.value = value!;
-                },
-                initialValue: controller.isGalleryRead.value,
-                title: RichText(
-                  text: const TextSpan(
-                    children: [
-                      TextSpan(
-                        text: 'Saya sudah melihat berkas ini',
-                        style: TextStyle(color: Colors.black),
-                      ),
-                    ],
-                  ),
-                ),
-                validator: FormBuilderValidators.equal(
-                  true,
-                  errorText: 'Saya sudah melihat berkas ini',
-                ),
-              ),
-            ],
           ),
         ),
       ),
