+++ lib/app/modules/reviewer_submit/widget/components/reviewer_submit_usaha/reviewer_submit_usaha_button.dart
@@ -133,11 +133,7 @@ class ReviewerSubmitUsahaButton extends StatelessWidget {
                     value: false,
                     child: Text(
                       'TIDAK',
-                      style: promptTextSubtitle(
-                          controller.isAnalisisUsahaRead.value
-                              ? Colors.transparent
-                              : Colors.grey[400]!,
-                          context),
+                      style: promptTextSubtitle(Colors.transparent, context),
                     ),
                   ),
                 ],
